+++ slosh_lib/Cargo.toml
@@ -1,17 +1,13 @@
 [package]
-name = "slosh"
+name = "slosh_lib"
 version = "0.10.0"
 edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[features]
-lisp-test = ["dep:regex", "dep:lazy_static"]
 
 [dependencies]
 unicode_reader = "1"
 sl-compiler = { workspace = true }
 bridge_adapters = { path = "../bridge_adapters" }
-slosh_lib = { path = "../slosh_lib" }
 sl-liner = { git = "https://github.com/sl-sh-dev/sl-liner.git" }
 slvm = { workspace = true }
 builtins = { path = "../builtins" }
@@ -20,10 +16,6 @@ shell = { path = "../shell" }
 unicode-width = "0.1"
 glob = "0.3"
 
-# lisp-test dependencies
-regex = { workspace = true, optional = true }
-lazy_static = { workspace = true, optional = true }
-
 [dev-dependencies]
 tempdir = "0.3.7"
 temp-env = "0.3.6"
@@ -31,16 +23,7 @@ regex = { workspace = true }
 lazy_static = { workspace = true }
 criterion = "0.5"
 iai-callgrind = "0.10.2"
-bench_utils = { path = "bench_utils" }
 compiler_test_utils = { workspace = true }
 
 [build-dependencies]
 chrono = "0.4.7"
-
-[[bench]]
-name = "criterion"
-harness = false
-
-[[bench]]
-name = "iai"
-harness = false
