+++ slosh_lib/src/completions.rs
